% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_users_uncertainty.R, R/run_subnational.R
\name{adjust_users_uncertainty}
\alias{adjust_users_uncertainty}
\alias{run_subnational}
\title{Adjusting quantity data to reflect users with uncertainty.}
\usage{
adjust_users_uncertainty(
  sectors_reporting_input,
  my_country_set_up,
  reporting_rates_df,
  baseline_users,
  commodities_table,
  method_continuation_df,
  user_input_adjustment_table
)

run_subnational(subnational_tools_filepath, input_shiny_type)
}
\arguments{
\item{sectors_reporting_input}{A table of values indicating if a sector is fully/partial/non-reporting.}

\item{my_country_set_up}{A table of primary country information.}

\item{reporting_rates_df}{A table of reporting rates for your chosen ss_type}

\item{baseline_users}{A table of baseline users calculated for the years prior to your chosen ss_type data.}

\item{commodities_table}{A table of commodities for your chosen service statistic type}

\item{method_continuation_df}{A table of method continuation rates for long term methods.}

\item{user_input_adjustment_table}{A table indicating whether a method should be adjusted or not.}

\item{subnational_tools_filepath}{The filepath in which all individual subnational SS-to-EMU tools are located.}

\item{input_shiny_type}{The service statistics type to use.}
}
\value{
A table returning the samples of users for each method each year adjusted for the missing private sector.

a list containing an emu database and a list of plots.
}
\description{
Adjusting quantity data to reflect users with uncertainty.

Running subnational shinyss2emu for subnational levels. Saves a CSV file and a pdf in the specified directory.
}
