% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_ss_to_emu.R
\name{country_ss_to_emu}
\alias{country_ss_to_emu}
\title{Calculating EMU from excel spreadsheet using modelled supply-share estimates, and saving output in specified files}
\usage{
country_ss_to_emu(
  country_tools_info,
  incl_condoms = FALSE,
  method_summary = FALSE
)
}
\arguments{
\item{country_tools_info}{Country tools info (get_tools_info() output)}

\item{incl_condoms}{include condoms in EMU output or not}
}
\description{
Calculating EMU from excel spreadsheet using modelled supply-share estimates, and saving output in specified files
}
